requirements.txt
azure-functions
python-dotenv
openai
Git-Python